+++ lib/drivers/dsn.go
@@ -1,4 +1,4 @@
-package main
+package drivers
 
 import (
 	"fmt"
