--- lib/test/file_ext_test/revisions/test.yaml
@@ -1,7 +1,0 @@
----
-- revision: revision_1
-  migrate:
-    - migrate 1
-    - migrate 2
-  rollback: 
-    - rollback 1
