--- lib/test/file_ext_test/revisions/test2.yml
@@ -1,7 +1,0 @@
----
-- revision: revision_2
-  migrate:
-    - migrate 1
-    - migrate 2
-  rollback: 
-    - rollback 1
