--- lib/test/parse_test_dir/config.yml
@@ -1,7 +1,0 @@
----
-type: mysql
-database: blog_db
-username: root
-password: root
-host: example.coom
-port: "3306"
