--- lib/test/parse_test_dir/revisions.yml
@@ -1,5 +1,0 @@
----
-revisions:
-  - revision_1
-  - revision_2
-
