--- lib/test/parse_test_dir_yaml_revisions/revisions.yaml
@@ -1,5 +1,0 @@
----
-revisions:
-  - revision_1
-  - revision_2
-
