--- lib/test/parse_test_dir_yaml_revisions/revisions/test_array.yaml
@@ -1,13 +1,0 @@
----
-- revision: revision_1
-  migrate:
-    - migrate 1
-  rollback: 
-    - rollback 1 
-- revision: revision_2 
-  migrate:
-    - migrate 1 
-  rollback:
-    - rollback 1
-    - rollback 2
-
